+++ src/run/constants.py
@@ -168,6 +168,6 @@ SEVEN_ZIP_PATH = read_ini_config(ini_file_path, "unzip_config", "SEVEN_ZIP_PATH"
 # unzip zip file password
 PASSWORD = read_ini_config(ini_file_path, "unzip_config", "PASSWORD")
 
-sis_server_version = "v1.1.5-beta.1.240508"
-build_date = "2024-05-08 18:00"
-publish_date = "2024-05-08 18:30"
+sis_server_version = "v1.1.5-beta.2.240513"
+build_date = "2024-05-13 18:00"
+publish_date = "2024-05-13 18:30"