+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## v1.1.5-beta.2.240513
+### 2024/5/13
+- add reduce memory usage to system_monitor
+
 ## v1.1.5-beta.1.240508
 ### 2024/5/8
 - catch exception for auto_image_explore.